修改。删除图片，判断是否有删除的权限

+++ app/libraries/class/Upload.php
@@ -113,12 +113,16 @@ class Upload
         $path = $this->filepath.'/'.$path;
         if( Storage::disk(self::$disk)->exists($path) )
         {
-            $images = Storage::disk(self::$disk)->delete( $path );
-            if ( $images )
-            {
+            if (is_writable($path)) {
+                $images = Storage::disk(self::$disk)->delete( $path );
+                if ( $images )
+                {
+                    return true;
+                }
+                return false;
+            } else {
                 return true;
             }
-            return false;
         }else
             return false;
     }
